restaurant_management_system/requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-dotenv
uvicorn
